created inital features structure / plan
Also added a small configuration fix to the tsconfig.vitest.json.
Fixed app title name in index.html to PhoneBook app.. as the name is
suggested in the README.md.

Initial types for the app will be in /features/index.ts
Person interface is based on API response, while the PhoneBookAddress is
a convenience interface